Remove vectors from GUI
The direction vectors (x, y) passed to the turn function were tightly
coupling the GUI to the logic. Replaced them with distinct turn_*
functions that are then handled in the logic.

Used some enums to pass too many arguments around and currying to create
the child functions from the generic one.