Manually trigger state change
As a first way to remove the state as a global variable, introduced a
new state_changed() trigger function. The point is that the logic
announces when the state changes and the GUI then updates the sprites.
It should result in a state when the game state is not mutable and the
logic returns the new state, which is then reflected in the GUI.