Make the event function return the new state
Replaced the state_changed trigger returning the new state from the
event handlers. This is still a bit fake, since the State instance is
mutable and that the old state passed to the events is still the
original instance passed on initialization and not the one returned.

Still a lot of work to do remains.